add a golden record mock factory class
