+++ bpdm-gate/src/test/kotlin/org/eclipse/tractusx/bpdm/gate/config/OrchestratorMockConfig.kt
@@ -38,46 +38,5 @@ import java.time.Duration
 @Configuration
 class OrchestratorMockConfig {
 
-    @Bean
-    fun refinementTestDataFactory(): RefinementTestDataFactory {
-        return RefinementTestDataFactory()
-    }
-
-    @Bean
-    fun orchestratorMockDataFactory(
-        refinementTestDataFactory: RefinementTestDataFactory,
-        requestFactory: OrchestratorRequestFactoryV7,
-        resultFactory: OrchestratorExpectedResultFactoryV7,
-        jsonMapper: JsonMapper
-    ): OrchestratorMockDataFactory {
-        return OrchestratorMockDataFactory(refinementTestDataFactory,requestFactory, resultFactory, jsonMapper)
-    }
-
-    @Bean
-    fun orchestratorRequestFactoryV7(testMetadataV7: TestMetadataV7): OrchestratorRequestFactoryV7{
-        val testMetadataReferences = TestMetadataReferences(
-            testMetadataV7.legalForms.map { it.technicalKey },
-            testMetadataV7.legalEntityIdentifierTypes.map { it.technicalKey },
-            testMetadataV7.addressIdentifierTypes.map { it.technicalKey },
-            testMetadataV7.adminAreas.map { it.code },
-            testMetadataV7.reasonCodes.map { it.technicalKey },
-            testMetadataV7.scriptCodes.map { it.technicalKey }
-        )
-
-        val orchestratorCommonFactory = OrchestratorRequestFactoryCommon(testMetadataReferences)
-        return OrchestratorRequestFactoryV7(BusinessPartnerTestDataFactory(orchestratorCommonFactory), orchestratorCommonFactory)
-    }
-
-    @Bean
-    fun orchestratorExpectedResultFactoryV7(): OrchestratorExpectedResultFactoryV7{
-        return OrchestratorExpectedResultFactoryV7(
-            Duration.ofDays(1),
-            Duration.ofDays(1),
-            mapOf(
-                Pair(TaskMode.UpdateFromPool, listOf(TaskStep.CleanAndSync, TaskStep.PoolSync)),
-                Pair(TaskMode.UpdateFromSharingMember, listOf(TaskStep.Clean))
-            )
-        )
-    }
 
 }
