+++ bpdm-gate/src/test/kotlin/org/eclipse/tractusx/bpdm/gate/v7/GateTestBaseV7.kt
@@ -31,6 +31,7 @@ import org.eclipse.tractusx.bpdm.gate.v7.util.GateTestClientProviderV7
 import org.eclipse.tractusx.bpdm.gate.v7.util.GateTestDataClientV7
 import org.eclipse.tractusx.bpdm.test.containers.KeyCloakInitializer
 import org.eclipse.tractusx.bpdm.test.testdata.gate.v7.TestDataFactoryGateV7
+import org.eclipse.tractusx.bpdm.test.testdata.GoldenRecordMockFactory
 import org.eclipse.tractusx.bpdm.test.testdata.orchestrator.OrchestratorMockDataFactory
 import org.eclipse.tractusx.bpdm.test.testdata.orchestrator.OrchestratorRequestFactoryV7
 import org.eclipse.tractusx.bpdm.test.testdata.pool.PoolMockDataFactory
@@ -63,6 +64,8 @@ abstract class GateTestBaseV7 : GateTestBase(){
     lateinit var orchestratorRequestFactoryV7: OrchestratorRequestFactoryV7
     @Autowired
     lateinit var poolMockDataFactory: PoolMockDataFactory
+    @Autowired
+    lateinit var goldenRecordMockFactory: GoldenRecordMockFactory
 
     lateinit var gateClient: GateClient
     lateinit var testDataClient: GateTestDataClientV7
@@ -78,7 +81,7 @@ abstract class GateTestBaseV7 : GateTestBase(){
             taskResolutionBatchService,
             relationTaskResolutionService,
             relationTaskCreationService,
-            poolMockDataFactory,
+            goldenRecordMockFactory,
             tenantBPNL
         )
     }
